1044596: Don't match beta product tags for release
Change the release module to reuse RHELProductMatcher
so it doesn't incorrectly identify content as being
RHEL content, so we don't attempt to find release
listings for them.

Add a rhelproduct.py with some of the RHEL product
specific tag comparisions, and update release
and productid modules to use it.
