+++ subscription-manager.spec
@@ -247,6 +247,7 @@ rm -rf %{buildroot}
 %{_datadir}/rhsm/subscription_manager/release.py*
 %{_datadir}/rhsm/subscription_manager/repolib.py*
 %{_datadir}/rhsm/subscription_manager/rhelentbranding.py*
+%{_datadir}/rhsm/subscription_manager/rhelproduct.py*
 %{_datadir}/rhsm/subscription_manager/utils.py*
 %{_datadir}/rhsm/subscription_manager/printing_utils.py*
 %{_datadir}/rhsm/subscription_manager/validity.py*
